rearranged static files and created template directory
