--- login_register/templates/login_view.html
@@ -1,5 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-<div>Login view</div>
-{% endblock %}
