--- login_register/templates/register_view.html
@@ -1,5 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-<div>Register view</div>
-{% endblock %}
