+++ login_register/views.py
@@ -1,7 +1,7 @@
 from django.shortcuts import render
 
 def login_view(request):
-    return render(request, 'login_view.html')
+    return render(request, 'loginandRegister/login_view.html')
 
 def register_view(request):
-    return render(request, 'register_view.html')
+    return render(request, 'loginandRegister/register_view.html')
