+++ templates/loginandRegister/login_view.html
@@ -1,0 +1,5 @@
+{% extends 'base.html' %}
+{% load static %}
+{% block content %}
+<div> <h1 class="h1"> Login view</h1></div>
+{% endblock content %}
