+++ templates/loginandRegister/register_view.html
@@ -1,0 +1,5 @@
+{% extends 'base.html' %}
+{% load static %}
+{% block content %}
+<div> <h2 class="h2"> Register view</h2></div>
+{% endblock content %}
