+++ templates/tetrisgame/game_view.html
@@ -1,6 +1,6 @@
 {% extends 'base.html' %}
 
-
+{% load static %}
 {% block content %}
 
 {% load static %}
@@ -15,4 +15,4 @@
         </div>
     </div>
 </div>
-{% endblock %}
+{% endblock content %}
