+++ tetris/static/js/tetris.js
