+++ tetris_app/views.py
@@ -1,4 +1,4 @@
 from django.shortcuts import render
 
 def game_view(request):
-    return render(request, 'game_view.html')
+    return render(request,'tetrisgame/game_view.html')
